+++ packages/core/src/components/progress-indicator/progress-indicator.template.ts
@@ -2,10 +2,10 @@ import { ProgressIndicator } from '@dso-toolkit/sources';
 import { html } from 'lit-html';
 import { ifDefined } from 'lit-html/directives/if-defined';
 
-export function progressIndicatorTemplate({ status, block, size }: ProgressIndicator) {
+export function progressIndicatorTemplate({ label, block, size }: ProgressIndicator) {
   return html`
     <dso-progress-indicator
-      status=${ifDefined(status)}
+      label=${ifDefined(label)}
       size=${ifDefined(size)}
       ?block=${block}
     ></dso-progress-indicator>