+++ packages/react/src/components/progress-indicator/progress-indicator.template.tsx
@@ -3,12 +3,12 @@ import * as React from 'react';
 
 import { DsoProgressIndicator } from '../..';
 
-export function progressIndicatorTemplate({ status, block, size }: ProgressIndicator) {
-  console.log(status, block, size);
-  
+export function progressIndicatorTemplate({ label, block, size }: ProgressIndicator) {
+  console.log(label, block, size);
+
   return (
     <DsoProgressIndicator
-      status={status}
+      label={label}
       block={block || undefined}
       size={size}
     />